requirements.txt
blinker==1.4
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytest==6.0.1
pytest-cov==2.10.0
pytest-mock==1.12.0
pyyaml==5.3.1
requests==2.24.0
Werkzeug==1.0.1